La til dokumentasjon og refaktorerte Trips.
La også til validator for tiden.
